env 서빙 클래스 만드는중, async db연결로 변경
