+++ main-server/src/app.module.ts
@@ -3,10 +3,10 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { MongooseModule } from '@nestjs/mongoose';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { getMongoUri } from './config/db/mongoose.config';
-import { getTypeormConfig } from './config/db/typeorm.config';
 import { ConfigModule } from '@nestjs/config';
 import { getEnvPath, validateDevEnv } from './config/env/dev-env.validation';
+import { TypeOrmConfigService } from './config/db/typeorm-config.service';
+import { MongooseConfigService } from './config/db/mongoose-config.service';
 
 @Module({
   imports: [
@@ -15,8 +15,8 @@ import { getEnvPath, validateDevEnv } from './config/env/dev-env.validation';
       isGlobal: true,
       validate: validateDevEnv,
     }),
-    TypeOrmModule.forRoot(getTypeormConfig()),
-    MongooseModule.forRoot(getMongoUri()),
+    TypeOrmModule.forRootAsync({ useClass: TypeOrmConfigService }),
+    MongooseModule.forRootAsync({ useClass: MongooseConfigService }),
   ],
   controllers: [AppController],
   providers: [AppService],
