--- main-server/src/config/db/mongoose.config.ts
@@ -1,7 +1,0 @@
-export const getMongoUri = () => {
-  // if (process.env.NODE_ENV === 'development') {
-  // }
-  return devMongoUri;
-};
-
-const devMongoUri = 'mongodb://mgdb:27017/aigoo';
