--- main-server/src/config/db/typeorm.config.ts
@@ -1,20 +1,0 @@
-import { TypeOrmModuleOptions } from '@nestjs/typeorm';
-import { SnakeNamingStrategy } from 'typeorm-naming-strategies';
-
-export const devTypeORMConfig: TypeOrmModuleOptions = {
-  type: 'postgres',
-  host: 'pgdb',
-  port: 5432,
-  username: 'postgres',
-  password: 'postgres',
-  database: 'postgres',
-  entities: [__dirname + '/../**/*.entity{.ts,.js}'],
-  synchronize: true,
-  namingStrategy: new SnakeNamingStrategy(),
-};
-
-export const getTypeormConfig = () => {
-  // if (process.env.NODE_ENV === 'development') {
-  // }
-  return devTypeORMConfig;
-};
